+++ .env.example
@@ -1,5 +1,5 @@
 POSTGRES_DB=edr
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=password
-POSTGRES_TABLE=cpnr
+POSTGRES_TABLE=resviz
 POSTGRES_HOST=localhost
