+++ docker-compose.yml
@@ -28,7 +28,7 @@ services:
     ports:
       - 5432:5432
     volumes:
-      - ./packages/cpnr/schema.sql:/docker-entrypoint-initdb.d/schema.sql
+      - ./packages/resviz/schema.sql:/docker-entrypoint-initdb.d/schema.sql
 
   redis:
     image: redislabs/redismod
