+++ packages/resviz/pyproject.toml
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: MIT
 
 [project]
-name = "cpnr"
+name = "resviz"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
@@ -21,10 +21,10 @@ requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
 
 [project.scripts]
-cpnr = "cpnr:load"
+resviz = "resviz:load"
 
 [tool.hatch.build.targets.wheel]
-packages = ["cpnr"]
+packages = ["resviz"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
